Moved  AMP detail drop-down to Header for fixed position. This commit contains changes  for restoring scalastyle xml to original definition. Removed scalariform plugin from build definition.
