+++ build.sbt
@@ -20,7 +20,7 @@ lazy val elideOptions = settingKey[Seq[String]]("Set limit for elidable function
 // instantiate the JS project for SBT with some additional settings
 // "client" for livelygig or "sclient" for synereo below
 // and edit server/src/main/twirl/views/index.scala.html file appropirately
-lazy val pCompile = "client"
+lazy val pCompile = "sclient"
 lazy val lessFile = if (pCompile == "sclient") "synereo-main.less" else "main.less"
 lazy val client: Project = (project in file(pCompile))
   .settings(
