+++ project/plugins.sbt
@@ -17,7 +17,7 @@ addSbtPlugin("com.typesafe.sbt" % "sbt-gzip" % "1.0.0")
 
 addSbtPlugin("org.scalastyle" %% "scalastyle-sbt-plugin" % "0.8.0")
 
-addSbtPlugin("org.scalariform" % "sbt-scalariform" % "1.6.0")
+//addSbtPlugin("org.scalariform" % "sbt-scalariform" % "1.6.0")
 
 //addSbtPlugin("com.typesafe.sbt" % "sbt-start-script" % "0.10.0")
 //addSbtPlugin("net.virtual-void" % "sbt-dependency-graph" % "0.8.2")
