+++ server/src/main/assets/stylesheets/synereo-main.less
@@ -331,7 +331,7 @@ label:hover:before {
   background: rgba(0, 0, 0, 0.5);
   z-index: 10;
   //transition: all 1s ease-in-out;
-  right: 0px;
+  right: 3px;
 }
 .topbar-left {
   left: 50;
@@ -537,6 +537,10 @@ transform: translate(0,0); */
 }
 #homeFeedMediaList li {
   opacity:0.2;
+}
+#homeFeedMediaList li:hover {
+  opacity:1 !important;
+  transform: scale(1.05);
 }
 /*
  * selectize.css
