+++ server/src/main/twirl/views/index.scala.html
@@ -6,13 +6,13 @@
         <meta charset="UTF-8">
         <title>@title</title>
         <meta content='width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no' name='viewport'>
-        <link rel="stylesheet" media="screen" href="@routes.Assets.at("/public","stylesheets/main.min.css")">
-        <link rel="shortcut icon" type="image/png" href="@routes.Assets.at("/public","images/favicon.ico")">
-        <script src="@routes.Assets.at("/public","javascripts/prolog_parser.js")" type="text/javascript"></script>
+        <!--<link rel="stylesheet" media="screen" href="@routes.Assets.at("/public","stylesheets/main.min.css")">-->
+        <!--<link rel="shortcut icon" type="image/png" href="@routes.Assets.at("/public","images/favicon.ico")">-->
+        <!--<script src="@routes.Assets.at("/public","javascripts/prolog_parser.js")" type="text/javascript"></script>-->
         <!--Synereo Specific links and scripts-->
-        <!--<link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
+        <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
         <link rel="stylesheet" media="screen" href="@routes.Assets.at("/public","stylesheets/synereo-main.min.css")">
-        <link rel="shortcut icon" type="image/png" href="@routes.Assets.at("/public","synereo-images/faviconSynereo.ico")">-->
+        <link rel="shortcut icon" type="image/png" href="@routes.Assets.at("/public","synereo-images/faviconSynereo.ico")">
 
     </head>
     <body>
